Bank card operations
There are 2 entries, customer and operator.
The user applies for a new card.
If the request is confirmed by the operator, the card number and pin will be sent to the e-mail address entered by the customer. Only then each user will be able to log in to their account by entering their pin code and 16-digit cart number.
After the operation, the updated information about the balance will be displayed on the screen. The user can change the pin code and track the entire history of income and expenditure in the form of a table with the exact date.
The Operator will also access the Operator's page with a predefined personal PIN code. This will allow him to obtain all the card information that has been applied for, approved, deleted or waiting for approval for the new card. He can also create a new card by entering the customer's information. In addition, card information may change.